refactoring - move parsing functions from cbird to syntax
